pyproject.toml
[project]
name = "repo-context"
version = "0.3.0"
description = "Convert Git repositories into LLM-friendly context format"
authors = [{ name = "Mathias Nielsen", email = "mathiasesn1@gmail.com" }]
maintainers = [{ name = "Mathias Nielsen", email = "mathiasesn1@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "gitpython>=3.1.43",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "tqdm>=4.67.1",
]

[tool.uv]
dev-dependencies = [
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
]

[project.scripts]
repo-context = "repo_context.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
